+++ packages/admin-ui-form/src/form/form.state.ts
@@ -1,7 +1,7 @@
 import type { RegisterOptions, UseFormReturn } from 'react-hook-form'
-import { useForm as useFormState } from 'react-hook-form'
+import { useForm as useFormState, useFieldArray } from 'react-hook-form'
 
 export type { RegisterOptions }
-export { useFormState }
+export { useFormState, useFieldArray }
 
 export type FormState = UseFormReturn<any>