+++ packages/admin-ui/src/table/components/styles/table.styles.ts
@@ -1,5 +1,5 @@
 import { style } from '@vtex/admin-ui-core'
-import type { TableColumn } from '../types'
+import type { TableColumn } from '../../types'
 
 interface RowOptions {
   columns: Array<TableColumn<any>>
@@ -19,7 +19,6 @@ const getColumnWidth = (column: TableColumn<any>) => {
 
 export const baseline = ({ columns }: RowOptions) =>
   style({
-    position: 'relative',
     display: 'grid',
     overflow: 'auto',
     gridTemplateColumns: columns.map(getColumnWidth).join(' '),