+++ packages/admin-ui-docs/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.7-dev.8 (2022-10-11)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## [0.16.7-dev.7](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-docs@0.16.7-dev.6...@vtex/admin-ui-docs@0.16.7-dev.7) (2022-10-07)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## 0.16.7-dev.6 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filtergroup:** add space to clear button ([07a9d93](https://github.com/vtex/admin-ui/commit/07a9d936de8a5b1d8cde725251d12c12a9506323))
+
+
+
+
+
+## [0.16.7-dev.5](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-docs@0.16.7-dev.4...@vtex/admin-ui-docs@0.16.7-dev.5) (2022-10-06)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## 0.16.7-dev.4 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filter popover:** fix internal spacing ([201e3be](https://github.com/vtex/admin-ui/commit/201e3be897ecaaccc410b7e83815fddf43135b39))
+
+
+
+
+
+## 0.16.7-dev.3 (2022-09-20)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
 ## [0.16.7-dev.2](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-docs@0.16.7-dev.1...@vtex/admin-ui-docs@0.16.7-dev.2) (2022-09-16)
 
 **Note:** Version bump only for package @vtex/admin-ui-docs