+++ packages/admin-ui-form/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.4.3-dev.7 (2022-10-11)
+
+**Note:** Version bump only for package @vtex/admin-ui-form
+
+
+
+
+
+## [0.4.3-dev.6](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-form@0.4.3-dev.5...@vtex/admin-ui-form@0.4.3-dev.6) (2022-10-07)
+
+
+### Features
+
+* **form:** export useFieldArray to improve the validation step ([0e62b7e](https://github.com/vtex/admin-ui/commit/0e62b7e3d8adcc69989de622b01c01f20bc8bf09))
+
+
+
+
+
+## [0.4.3-dev.5](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-form@0.4.3-dev.4...@vtex/admin-ui-form@0.4.3-dev.5) (2022-10-07)
+
+**Note:** Version bump only for package @vtex/admin-ui-form
+
+
+
+
+
+## 0.4.3-dev.4 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filtergroup:** add space to clear button ([07a9d93](https://github.com/vtex/admin-ui/commit/07a9d936de8a5b1d8cde725251d12c12a9506323))
+
+
+
+
+
+## 0.4.3-dev.3 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filter popover:** fix internal spacing ([201e3be](https://github.com/vtex/admin-ui/commit/201e3be897ecaaccc410b7e83815fddf43135b39))
+
+
+
+
+
+## 0.4.3-dev.2 (2022-09-20)
+
+**Note:** Version bump only for package @vtex/admin-ui-form
+
+
+
+
+
 ## 0.4.3-dev.1 (2022-09-15)
 
 **Note:** Version bump only for package @vtex/admin-ui-form