+++ packages/admin-ui-react/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.9.5-dev.5 (2022-10-11)
+
+**Note:** Version bump only for package @vtex/admin-ui-react
+
+
+
+
+
+## 0.9.5-dev.4 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filtergroup:** add space to clear button ([07a9d93](https://github.com/vtex/admin-ui/commit/07a9d936de8a5b1d8cde725251d12c12a9506323))
+
+
+
+
+
+## 0.9.5-dev.3 (2022-10-06)
+
+
+### Bug Fixes
+
+* **filter popover:** fix internal spacing ([201e3be](https://github.com/vtex/admin-ui/commit/201e3be897ecaaccc410b7e83815fddf43135b39))
+
+
+
+
+
+## 0.9.5-dev.2 (2022-09-20)
+
+**Note:** Version bump only for package @vtex/admin-ui-react
+
+
+
+
+
 ## 0.9.5-dev.1 (2022-09-15)
 
 **Note:** Version bump only for package @vtex/admin-ui-react